+++ src/main/java/org/iti/ecomus/entity/Product.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -16,7 +17,7 @@ public class Product implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "productId")
-    private int productId;
+    private Long productId;
 
     @Column(name = "productName", nullable = false)
     @NotEmpty
@@ -31,14 +32,14 @@ public class Product implements Serializable {
 
     @Column(name = "price", nullable = false)
     @Min(1)
-    private int price;
+    private BigDecimal price;
 
     @ManyToMany(mappedBy = "products") // "products" is the name of the field in Category
     private List<Category> categories;
 
     public Product() {}
 
-    public Product(String productName, String description, int quantity, int price) {
+    public Product(String productName, String description, int quantity, BigDecimal price) {
         this.productName = productName;
         this.description = description;
         this.quantity = quantity;
